Add good suffix table generation for case 1, shifting according to matching sufffix
This is only calculating the starting position of the widest border of the suffix
of the pattern beginning at the position i.
Take for instance the pattern: abbabab
If we start at index 0, we calculate the widest border (A border is a
substring which is both a proper prefix and a proper suffix of a string)
and that would be ab, which starts at position 5, if we move to index
1, we have: bbabab, and the widest border for that would be b, so its
index value would be 6. When moving to index 2, with string: babab, the
widest border would be bab, which starts at 4, and so on. Note that we
when we are at index 6, b has no border, so we equat it to length + 1,
i.e. 7

 i:  0 1 2 3 4 5 6 7
 p:  a b b a b a b
 b:  5 6 4 5 6 7 7 8

 As for the actual implementation, we start from right to left, and we
 have two cursors the i, which iterates over the indices of the string,
 and j which starts with the value: length + 1, and helps us compare characters
 across the string. J starts at index of the smallest border, and when
 we compare the characters that lie in i with j, if there is a mismatch
 we shift j to the wider border of the string, until we get a match
 between the characters, we are able to determine the largest border
 then, and set our value to j. These calculations are not actually used
 as shifts, but help us calculate the shift we need, to be implemented
 in next commit