Add shift to good suffix table
The shift is calculated according to the borders that cannot be extended to the
left. This happens three times in the abbabab, twice at index 6, where
there is a mismatch but we default to the first value, and the second
time at index 4. The rest of the values default to zero, please note
this is one of the good suffix shift tables that satisfies the matching
suffix occuring somewhere else in the pattern